Adding Directory for a Basic Golang Webserver
Three Routes are implemented (root, hello, form) all implemented in pure html. The root contains index.html file that is the landing page. 'hello' is the first function that prints 'hello' when correct route that is /hello is entered in URL entry. 'form' contains the form.html that parses User Data (Name and Address) and "is not storing/pushing the data in some local/cloud storage" till the date.